requirements.txt
discord.py>=2.3
apscheduler>=3.10
rapidfuzz>=3.0
requests>=2.31
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.21
